src/backend/requirements.txt
anyio==3.6.2
appdata==2.1.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.6
colored==1.4.3
cryptography==38.0.3
DateTime==4.7
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.87.0
gcloud==0.18.3
googleapis-common-protos==1.57.0
h11==0.12.0
httpcore==0.15.0
httplib2==0.21.0
httptools==0.5.0
httpx==0.23.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
jwcrypto==1.4.2
MarkupSafe==2.1.1
oauth2client==4.1.3
orjson==3.8.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dotenv==0.21.0
python-jose==3.3.0
python-jwt==4.0.0
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
to-requirements.txt==1.3.0
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.19.0
watchfiles==0.18.1
websockets==10.4
wrapt==1.14.1
zope.interface==5.5.1
